+++ src/redux/saga/todo/addTodo.js
@@ -1,11 +1,11 @@
 import { takeLatest, call, put } from 'redux-saga/effects'
 
-import { TODO_ADD_REQUESTED } from '../types/todoTypes'
-import API from '../../api'
+import { TODO_ADD_REQUESTED } from '../../types/todoTypes'
+import API from '../../../api'
 import {
   addTodoActionFailure,
   addTodoActionSuccess
-} from '../actions/todo/addTodoAction'
+} from '../../actions/todo/addTodoAction'
 
 export function * addTodoSaga (action) {
   try {
